Refactor the activity instrumentation test cases
